refactor(pop): factor out request_queue
