+++ execution-engine/contracts/hdac-system/pop/src/lib.rs
@@ -10,7 +10,6 @@ mod contract_runtime;
 mod contract_stakes;
 mod contract_votes;
 mod pop_contract;
-mod request_queue;
 
 use alloc::string::String;
 
@@ -95,7 +94,8 @@ pub fn delegate() {
                 .finalize_payment(amount_spent, account)
                 .unwrap_or_revert();
         }
-        // Type of this method: `fn delegate(validator: PublicKey, amount: U512, src_purse_uref: URef)`
+        // Type of this method: `fn delegate(validator: PublicKey, amount: U512, src_purse_uref:
+        // URef)`
         methods::METHOD_DELEGATE => {
             let delegator = runtime::get_caller();
             let validator: PublicKey = runtime::get_arg(1)
@@ -124,7 +124,8 @@ pub fn delegate() {
                 .undelegate(delegator, validator, shares)
                 .unwrap_or_revert();
         }
-        // Type of this method: `fn redelegate(src_validator: PublicKey, dest_validator: PublicKey, amount: U512)`
+        // Type of this method: `fn redelegate(src_validator: PublicKey, dest_validator: PublicKey,
+        // amount: U512)`
         methods::METHOD_REDELEGATE => {
             let delegator: PublicKey = runtime::get_caller();
             let src_validator: PublicKey = runtime::get_arg(1)
